Fists & More Destructible Buildings
Buildings can now only be destroyed if the grip and trigger are both held down. The modified VRPawn stopped working since VRTemplate and its associated files are not in Git. To get around this, I've uploaded a zipped version of the modified VR stuff -- make sure to unzip from that rather than adding the VRTemplate in-editor.